implemented oauth token endpoint

+++ src/BeselerNet.Shared/Contracts/OAuthTokenResponse.cs
@@ -7,7 +7,7 @@ public sealed record class OAuthTokenResponse
     [JsonPropertyName("token_type")]
     public string TokenType { get; init; } = "Bearer";
     [JsonPropertyName("expires_in")]
-    public int ExpiresIn { get; init; }
+    public long ExpiresIn { get; init; }
     [JsonPropertyName("refresh_token")]
     public string? RefreshToken { get; init; }
 }
